// react-module/src/openapi/models/input-entity-with-placement.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface InputEntityWithPlacement
 */
export interface InputEntityWithPlacement {
    /**
     * 
     * @type {InputEntity}
     * @memberof InputEntityWithPlacement
     */
    inputEntity?: any;
    /**
     * 
     * @type {PlacedBoxMl}
     * @memberof InputEntityWithPlacement
     */
    placedBox?: any;
}