// react-module/src/openapi/models/rectangle-shape-extended-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface RectangleShapeExtendedDto
 */
export interface RectangleShapeExtendedDto {
    /**
     * 
     * @type {number}
     * @memberof RectangleShapeExtendedDto
     */
    id?: any;
    /**
     * 
     * @type {number}
     * @memberof RectangleShapeExtendedDto
     */
    height?: any;
    /**
     * 
     * @type {number}
     * @memberof RectangleShapeExtendedDto
     */
    width?: any;
}