// react-module/src/openapi/models/surface.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface Surface
 */
export interface Surface {
    /**
     * 
     * @type {number}
     * @memberof Surface
     */
    id?: any;
    /**
     * 
     * @type {number}
     * @memberof Surface
     */
    width?: any;
    /**
     * 
     * @type {number}
     * @memberof Surface
     */
    height?: any;
}