// react-module/src/openapi/models/surface-with-placed-entities.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface SurfaceWithPlacedEntities
 */
export interface SurfaceWithPlacedEntities {
    /**
     * 
     * @type {Surface}
     * @memberof SurfaceWithPlacedEntities
     */
    surface?: any;
    /**
     * 
     * @type {Array&lt;InputEntityWithPlacement&gt;}
     * @memberof SurfaceWithPlacedEntities
     */
    inputEntityWithPlacements?: any;
    /**
     * 
     * @type {number}
     * @memberof SurfaceWithPlacedEntities
     */
    fillRatio?: any;
}